Split AppleSDKPlatform, Platform into own files.
